refactor: 重构API响应实体类，统一结构并替换旧实现
本次提交对项目内大量API响应实体类进行了重构：
1. 新增了多个场景的响应实体类，包括群文件操作、语音相关、工具类接口等响应数据
2. 替换了旧版的响应实体类实现，统一字段命名和注解风格
3. 调整了部分实体类的包结构，将同类响应归类到对应包下
4. 更新了部分字段的类型和注释，提升代码可读性和规范性
5. 新增了scripts目录到gitignore中

+++ src/main/java/com/github/zhygtx/napcat/api/response/system/CsrfTokenData.java
@@ -3,13 +3,13 @@ package com.github.zhygtx.napcat.api.response.system;
 import lombok.Data;
 
 /**
- * CSRF Token 响应数据。
+ * 获取 CSRF Token
  * <p>
- * 对应 <code>/get_csrf_token</code> 的 data 字段。
  */
 @Data
 public class CsrfTokenData {
 
-    /** token 值 */
-    private int token;
+    /** CSRF Token */
+    private Long token;
+
 }
